requirements.txt
nltk==3.6.5
numpy==1.20.3
pandas==1.3.4
scikit_learn==1.0.2
streamlit==1.8.1
tensorflow==2.8.0
tensorflow_hub==0.12.0
xgboost
tensorflow_hub==0.12.0
h5py==3.1.0
opencv-python-headless
opencv-contrib-python-headless
Keras-Preprocessing==1.1.2
tensorflow-estimator==2.8.0
scipy==1.7.1
matplotlib==3.4.3
